分拆dayu.cli和dayu.wechat去掉lazy import
