+++ AGENTS.md
@@ -28,6 +28,8 @@
 - 函数必须提供完整中文 docstring，至少包含参数、返回值、异常。
 - 类与模块应提供中文概览 docstring；复杂逻辑必须补充中文行内注释说明意图。
 - 禁止使用 `object`、`Any`、无类型参数、无类型返回值，以及其他无法进行严格类型检查的签名设计。
+- 禁止胶水 seam，使用lazy import必须有充分理由。
+- 使用 `hasattr` 必须有充分理由，不能把它当作逃避类型与边界设计的手段。
 - 禁止把显式参数放进 `extra payload`。
 - 禁止魔法数字、魔法字符串；工具 schema 例外，schema 内允许直接写字面量字符串。
 - 优先使用模块级私有辅助函数；禁止无必要的嵌套函数、嵌套类。
@@ -40,7 +42,6 @@
   - 兼容性 wrapper / facade：方法体仅透传到真源模块，不增加有效语义。
 - 默认按全新设计处理，不为旧实现、旧接口、旧测试保留兼容逻辑。
 - 编写规则时优先自适应实现，禁止把业务规则硬编码成脆弱分支。
-- 使用 `hasattr` 必须有充分理由，不能把它当作逃避类型与边界设计的手段。
 - 涉及 schema 变更时，一律按全新 schema 起库处理；禁止旧库迁移、兼容读取、兼容测试，除非当前任务明确要求兼容升级。
 
 ## 测试与验证
