+++ dayu/cli/__init__.py
@@ -1,29 +1,6 @@
 """统一 CLI 入口包。"""
 
-from importlib import import_module
-from types import ModuleType
-from typing import TYPE_CHECKING
-
-if TYPE_CHECKING:
-    from dayu.cli import main as main
-
-
-def __getattr__(name: str) -> ModuleType:
-    """按需暴露 CLI 主入口模块。
-
-    Args:
-        name: 属性名。
-
-    Returns:
-        `dayu.cli.main` 模块对象。
-
-    Raises:
-        AttributeError: 当属性名不受支持时抛出。
-    """
-
-    if name != "main":
-        raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
-    return import_module("dayu.cli.main")
+import dayu.cli.main as main
 
 
 def run_main() -> int:
@@ -39,7 +16,7 @@ def run_main() -> int:
         无。
     """
 
-    return import_module("dayu.cli.main").main()
+    return main.main()
 
 
 __all__ = [
