+++ dayu/cli/commands/__init__.py
@@ -1,0 +1,7 @@
+"""`dayu.cli` 命令执行包。
+
+模块职责：
+- 承载各顶层 CLI 子命令的执行入口。
+- 保持 `dayu.cli.main` 仅负责轻量参数分发，不直接装配运行时。
+"""
+
