+++ dayu/wechat/commands/__init__.py
@@ -1,0 +1,1 @@
+"""WeChat CLI 命令执行包。"""
