+++ tests/README.md
@@ -171,8 +171,9 @@ python -m pytest tests --cov=dayu --cov-report=term --cov-branch
 - `test_write_pipeline.py` 还要守住证据锚点轻量修复的可观测性边界：`maybe_rewrite_evidence_anchors()` 即使在后验校验失败回退原文时，也必须把 `attempted/applied/failure_reason` 显式写入 `process_state.latest_anchor_rewrite` 与历史记录，不能再让调用方只能靠正文是否变化去猜测“未尝试”和“尝试失败”。
 - `test_write_pipeline.py` 还要守住 manifest 文件锁的跨平台边界：`artifact_store._manifest_file_lock()` 在 POSIX 必须继续走 `fcntl.flock()`，在 Windows 必须改用 `msvcrt.locking()` 提供真实互斥；其中 blocking 语义必须显式轮询直到拿到锁，不能退回 `LK_LOCK` 那种有限次重试；如果当前平台两种实现都不可用，也必须显式失败，不能静默跳过锁操作。
 - `test_write_pipeline.py` 还要守住审计模块拆分边界：`audit_formatting.py` 只放 Markdown 文本操作（标题/证据行/内容提取/patch 匹配），`audit_rules.py` 只放审计决策真源（解析/规则/修复合同/复核合并），`audit_evidence_rewriter.py` 只放证据锚点重写。三个模块之间的依赖方向是 `formatting ← rules` 和 `formatting ← evidence_rewriter → rules`，不允许反向。
-- `test_cli_running_config.py` 和 `test_cli_interactive_coverage.py` 还要守住 CLI 模块拆分边界：`arg_parsing.py` 只放 argparse 参数定义和解析器构建，`dependency_setup.py` 只放数据类型定义、配置解析和 Service 构建，`fins_commands.py` 只放财报命令构建和执行，`main.py` 只做命令分发。monkeypatch 路径必须指向被调用函数所在的实际模块命名空间（例如 `_build_fins_ops_service` 在 `fins_commands` 命名空间被调用，monkeypatch 应指向 `dayu.cli.fins_commands._build_fins_ops_service`），不允许通过包级 re-export 绕行。
-- `test_entrypoints.py` 还要守住 CLI / WeChat 的冷启动边界：导入 `dayu.cli.main` 时不得抢先导入 `host_commands`、`dependency_setup` 这类重运行时模块，导入 `dayu.wechat.main` 时也不得提前拉起 `dayu.wechat.daemon`；`--help` 与 `dayu-cli init` 这类薄入口必须继续保持 lazy import 结构。
+- `test_cli_running_config.py` 和 `test_cli_interactive_coverage.py` 还要守住 CLI 模块拆分边界：`arg_parsing.py` 只放 argparse 参数定义和解析器构建，`dependency_setup.py` 只放数据类型定义、配置解析和 Service 构建，`commands/` 承担各子命令执行，`main.py` 只做命令分发。monkeypatch 路径必须指向被调用函数所在的实际模块命名空间（例如 `_build_fins_ops_service` 在 `dayu.cli.commands.fins` 被调用，patch 应指向 `dayu.cli.commands.fins._build_fins_ops_service`），不允许通过包级 re-export 或旧 `main.py` seam 绕行。
+- `test_wechat_main.py` 还要守住 WeChat 模块拆分边界：`arg_parsing.py` 只负责参数解析与上下文收口，`runtime.py` 只负责 WeChat 运行时装配与 service/runtime helper，`commands/` 承担 `login / run / service` 子命令执行，`main.py` 只做命令分发。monkeypatch 必须 patch 到真实被调用模块（例如 `_has_persisted_wechat_login` 在 `dayu.wechat.commands.service` 被调用时，应 patch `dayu.wechat.commands.service._has_persisted_wechat_login`，不能回退去 patch 旧 `main.py` seam 或其他转发层）。
+- `test_entrypoints.py` 还要守住 CLI / WeChat 的冷启动边界：导入 `dayu.cli.main` 时不得抢先导入 `dayu.cli.commands.*`、`dependency_setup` 这类重运行时模块，导入 `dayu.wechat.main` 时也不得提前拉起 `dayu.wechat.daemon`；`--help` 与 `dayu-cli init` 这类薄入口必须继续保持“主入口只分发、命令模块按需导入”的结构。
 - `test_chat_service.py` 这类 Service/Host 边界测试里，如果断言必须观察 Host 内部 stub（如 `_executor`、`_run_registry`、`_pending_turn_store`），应先通过测试 helper 把 `ConversationalExecutionGatewayProtocol` 显式收窄到测试用具体 `Host` / stub 类型；不要把私有属性访问直接散落在测试正文里。
 - `tests/application/conftest.py` 这类共享夹具文件必须直接跟随 `HostExecutorProtocol` / `SessionRegistryProtocol` 演进；像 `StubHostExecutor.run_operation_stream/run_operation_sync` 这类泛型返回签名、以及 `StubSessionRegistry.close_idle_sessions()` 这类新协议方法，应优先在共享夹具真源补齐，而不是在每个下游测试里重复 `cast` 或 `type: ignore`。
 - 项目级 `tests/conftest.py` 这类 import 兼容夹具若需要预加载模块，也要先显式断言 `ModuleSpec/loader` 非空，再在 `ModuleType` 边界集中 `cast(Any, module)` 后补动态别名；不要把 `spec.loader` 的可空分支或动态属性赋值散落到正文。
