+++ tests/engine/test_cli_running_config.py
@@ -15,6 +15,9 @@ import pytest
 
 from dayu.contracts.cancellation import CancelledError
 from dayu.cli.arg_parsing import parse_arguments
+from dayu.cli.commands.interactive import run_interactive_command
+from dayu.cli.commands.prompt import run_prompt_command
+from dayu.cli.commands.write import run_write_command
 from dayu.cli.dependency_setup import (
     ModelName,
     RunningConfig,
@@ -32,7 +35,7 @@ from dayu.cli.dependency_setup import (
     setup_paths,
     setup_write_config,
 )
-from dayu.cli.fins_commands import _build_fins_command
+from dayu.cli.commands.fins import _build_fins_command, run_fins_command
 from dayu.cli.main import main
 from dayu.cli.interactive_state import (
     FileInteractiveStateStore,
@@ -670,13 +673,9 @@ def _raise_runtime_error_for_write_pipeline(**_kwargs: object) -> int:
 def test_main_dispatches_host_command(monkeypatch: pytest.MonkeyPatch, tmp_path: Path) -> None:
     """host 子命令应直接分发到宿主管理入口。"""
 
-    workspace = tmp_path / "workspace"
-    workspace.mkdir(parents=True, exist_ok=True)
-    monkeypatch.setattr(sys, "argv", ["prog", "host", "--base", str(workspace), "status"])
-    monkeypatch.setattr("dayu.cli.main.run_host_command", lambda args: 0)
-    monkeypatch.setattr("dayu.cli.main.setup_paths", lambda _args: pytest.fail("host 命令不应走 setup_paths"))
-    monkeypatch.setattr("dayu.cli.main.Log.info", lambda *_args, **_kwargs: None)
-    monkeypatch.setattr("dayu.cli.main.Log.error", lambda *_args, **_kwargs: None)
+    args = Namespace(command="host")
+    monkeypatch.setattr("dayu.cli.main.parse_arguments", partial(_return_value, args))
+    monkeypatch.setattr("dayu.cli.commands.host.run_host_command", lambda args: 0)
 
     assert main() == 0
 
@@ -685,12 +684,9 @@ def test_main_dispatches_host_command(monkeypatch: pytest.MonkeyPatch, tmp_path:
 def test_main_dispatches_init_without_runtime_setup(monkeypatch: pytest.MonkeyPatch, tmp_path: Path) -> None:
     """`init` 子命令不应触发运行时装配路径。"""
 
-    workspace = tmp_path / "workspace"
-    workspace.mkdir(parents=True, exist_ok=True)
-    monkeypatch.setattr(sys, "argv", ["prog", "init", "--base", str(workspace)])
-    monkeypatch.setattr("dayu.cli.main.run_init", lambda args: 7)
-    monkeypatch.setattr("dayu.cli.main.setup_loglevel", lambda _args: pytest.fail("init 命令不应设置日志"))
-    monkeypatch.setattr("dayu.cli.main.setup_paths", lambda _args: pytest.fail("init 命令不应解析运行时路径"))
+    args = Namespace(command="init")
+    monkeypatch.setattr("dayu.cli.main.parse_arguments", partial(_return_value, args))
+    monkeypatch.setattr("dayu.cli.commands.init.run_init_command", lambda args: 7)
 
     assert main() == 7
 
@@ -1678,7 +1674,7 @@ def test_setup_paths_warns_and_allows_missing_filings_dir_for_agent_commands(
     workspace_dir = tmp_path / "workspace"
     workspace_dir.mkdir(parents=True, exist_ok=True)
     collector = _CallCollector()
-    monkeypatch.setattr("dayu.cli.main.Log.warning", collector.capture_warn)
+    monkeypatch.setattr("dayu.cli.dependency_setup.Log.warning", collector.capture_warn)
 
     args = Namespace(base=str(workspace_dir), ticker="aapl", command=command_name)
     paths_config = setup_paths(args)
@@ -2007,6 +2003,103 @@ def test_setup_write_config_raises_system_exit_when_retry_count_is_negative(tmp_
 
 
 @pytest.mark.unit
+def test_setup_write_config_prefers_workspace_template_when_cli_not_provided(tmp_path: Path) -> None:
+    """验证未传 `--template` 时优先使用工作区模板。"""
+
+    workspace_dir = tmp_path / "workspace"
+    (workspace_dir / "config").mkdir(parents=True, exist_ok=True)
+    (workspace_dir / "assets").mkdir(parents=True, exist_ok=True)
+    (workspace_dir / "assets" / "定性分析模板.md").write_text("# workspace", encoding="utf-8")
+    (workspace_dir / "portfolio" / "AAPL" / "filings").mkdir(parents=True, exist_ok=True)
+    paths_config = WorkspaceConfig(
+        workspace_dir=workspace_dir,
+        output_dir=workspace_dir / "output",
+        config_loader=ConfigLoader(ConfigFileResolver(workspace_dir / "config")),
+        prompt_asset_store=FilePromptAssetStore(ConfigFileResolver(workspace_dir / "config")),
+        ticker="AAPL",
+        has_local_filings=True,
+    )
+    args = Namespace(
+        command="write",
+        output=None,
+        template=None,
+        write_max_retries=2,
+        resume=True,
+        web_provider=None,
+        audit_model_name=None,
+        chapter=None,
+        fast=False,
+        force=False,
+        infer=False,
+    )
+    running_config = RunningConfig(
+        runner_running_config=AsyncOpenAIRunnerRunningConfig(),
+        agent_running_config=AgentRunningConfig(),
+        doc_tool_limits=DocToolLimits(),
+        fins_tool_limits=FinsToolLimits(),
+        web_tools_config=WebToolsConfig(provider="auto"),
+        tool_trace_config=ToolTraceConfig(enabled=False, output_dir=tmp_path / "trace"),
+    )
+
+    write_config = setup_write_config(args, paths_config, running_config)
+
+    assert write_config.template_path == (workspace_dir / "assets" / "定性分析模板.md")
+
+
+@pytest.mark.unit
+def test_setup_write_config_falls_back_to_package_template_when_workspace_missing(
+    tmp_path: Path,
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    """验证工作区模板缺失时回退到包内默认模板。"""
+
+    workspace_dir = tmp_path / "workspace"
+    (workspace_dir / "config").mkdir(parents=True, exist_ok=True)
+    (workspace_dir / "portfolio" / "AAPL" / "filings").mkdir(parents=True, exist_ok=True)
+    package_assets_dir = tmp_path / "package_assets"
+    package_assets_dir.mkdir(parents=True, exist_ok=True)
+    (package_assets_dir / "定性分析模板.md").write_text("# package", encoding="utf-8")
+    monkeypatch.setattr(
+        "dayu.cli.dependency_setup.resolve_package_assets_path",
+        lambda: package_assets_dir,
+    )
+
+    paths_config = WorkspaceConfig(
+        workspace_dir=workspace_dir,
+        output_dir=workspace_dir / "output",
+        config_loader=ConfigLoader(ConfigFileResolver(workspace_dir / "config")),
+        prompt_asset_store=FilePromptAssetStore(ConfigFileResolver(workspace_dir / "config")),
+        ticker="AAPL",
+        has_local_filings=True,
+    )
+    args = Namespace(
+        command="write",
+        output=None,
+        template=None,
+        write_max_retries=2,
+        resume=True,
+        web_provider=None,
+        audit_model_name=None,
+        chapter=None,
+        fast=False,
+        force=False,
+        infer=False,
+    )
+    running_config = RunningConfig(
+        runner_running_config=AsyncOpenAIRunnerRunningConfig(),
+        agent_running_config=AgentRunningConfig(),
+        doc_tool_limits=DocToolLimits(),
+        fins_tool_limits=FinsToolLimits(),
+        web_tools_config=WebToolsConfig(provider="auto"),
+        tool_trace_config=ToolTraceConfig(enabled=False, output_dir=tmp_path / "trace"),
+    )
+
+    write_config = setup_write_config(args, paths_config, running_config)
+
+    assert write_config.template_path == (package_assets_dir / "定性分析模板.md")
+
+
+@pytest.mark.unit
 def test_resolve_write_output_dir_uses_ticker_subdir_by_default(tmp_path: Path) -> None:
     """验证写作输出目录默认落到 `draft/{ticker}`。
 
@@ -2092,7 +2185,7 @@ def test_setup_loglevel_selects_expected_level(monkeypatch: pytest.MonkeyPatch) 
     """
 
     collector = _CallCollector()
-    monkeypatch.setattr("dayu.cli.main.Log.set_level", collector.capture_level)
+    monkeypatch.setattr("dayu.cli.dependency_setup.Log.set_level", collector.capture_level)
 
     setup_loglevel(Namespace(log_level="info", debug=False, verbose=False, info=False, quiet=False))
     setup_loglevel(Namespace(log_level=None, debug=True, verbose=False, info=False, quiet=False))
@@ -2171,22 +2264,25 @@ def test_main_interactive_path_returns_zero(monkeypatch: pytest.MonkeyPatch, tmp
         collector.mark_interactive()
         interactive_kwargs.update(kwargs)
 
-    monkeypatch.setattr("dayu.cli.main.parse_arguments", partial(_return_value, args))
-    monkeypatch.setattr("dayu.cli.main.setup_paths", partial(_return_value, workspace_config))
-    monkeypatch.setattr("dayu.cli.main.setup_model_name", partial(_return_value, model_name))
+    monkeypatch.setattr("dayu.cli.commands.interactive.setup_loglevel", lambda _args: None)
+    monkeypatch.setattr("dayu.cli.commands.interactive.setup_paths", partial(_return_value, workspace_config))
+    monkeypatch.setattr(
+        "dayu.cli.commands.interactive._build_execution_options",
+        lambda _args: SimpleNamespace(model_name="deepseek-thinking"),
+    )
     fake_dependencies = _FakeCliHostDependencies(
         running_config=running_config,
         scene_model_names={"interactive": "scene-interactive-model"},
     )
     monkeypatch.setattr(
-        "dayu.cli.main._prepare_cli_host_dependencies",
+        "dayu.cli.commands.interactive._prepare_cli_host_dependencies",
         lambda **_kwargs: fake_dependencies.as_tuple(),
     )
-    monkeypatch.setattr("dayu.cli.main._build_chat_service", lambda **_kwargs: object())
-    monkeypatch.setattr("dayu.cli.main.Log.info", collector.capture_info)
-    monkeypatch.setattr("dayu.cli.main.interactive", _capture_interactive)
+    monkeypatch.setattr("dayu.cli.commands.interactive._build_chat_service", lambda **_kwargs: object())
+    monkeypatch.setattr("dayu.cli.commands.interactive.Log.info", collector.capture_info)
+    monkeypatch.setattr("dayu.cli.commands.interactive.interactive", _capture_interactive)
 
-    assert main() == 0
+    assert run_interactive_command(args) == 0
     assert collector.interactive_calls == 1
     state = FileInteractiveStateStore(tmp_path / ".dayu" / "interactive").load()
     assert state is not None
@@ -2222,12 +2318,14 @@ def test_main_interactive_path_rejects_second_instance(monkeypatch: pytest.Monke
     )
     error_logs: list[str] = []
 
-    monkeypatch.setattr("dayu.cli.main.parse_arguments", partial(_return_value, args))
-    monkeypatch.setattr("dayu.cli.main.setup_paths", partial(_return_value, workspace_config))
-    monkeypatch.setattr("dayu.cli.main.setup_model_name", partial(_return_value, ModelName(model_name="mimo-v2-flash")))
-    monkeypatch.setattr("dayu.cli.main._build_execution_options", lambda _args: None)
+    monkeypatch.setattr("dayu.cli.commands.interactive.setup_loglevel", lambda _args: None)
+    monkeypatch.setattr("dayu.cli.commands.interactive.setup_paths", partial(_return_value, workspace_config))
     monkeypatch.setattr(
-        "dayu.cli.main._prepare_cli_host_dependencies",
+        "dayu.cli.commands.interactive._build_execution_options",
+        lambda _args: SimpleNamespace(model_name="deepseek-thinking"),
+    )
+    monkeypatch.setattr(
+        "dayu.cli.commands.interactive._prepare_cli_host_dependencies",
         lambda **_kwargs: _FakeCliHostDependencies(
             running_config=RunningConfig(
                 runner_running_config=AsyncOpenAIRunnerRunningConfig(),
@@ -2240,15 +2338,15 @@ def test_main_interactive_path_rejects_second_instance(monkeypatch: pytest.Monke
             scene_model_names={"interactive": "scene-interactive-model"},
         ).as_tuple(),
     )
-    monkeypatch.setattr("dayu.cli.main._build_chat_service", lambda **_kwargs: object())
+    monkeypatch.setattr("dayu.cli.commands.interactive._build_chat_service", lambda **_kwargs: object())
     monkeypatch.setattr(
-        "dayu.cli.main.StateDirSingleInstanceLock.acquire",
+        "dayu.cli.commands.interactive.StateDirSingleInstanceLock.acquire",
         lambda self: (_ for _ in ()).throw(RuntimeError("同一个 state_dir 已有运行中的 interactive 单实例锁")),
     )
-    monkeypatch.setattr("dayu.cli.main.Log.error", lambda message, **_kwargs: error_logs.append(str(message)))
-    monkeypatch.setattr("dayu.cli.main.interactive", lambda *_args, **_kwargs: pytest.fail("不应进入 interactive"))
+    monkeypatch.setattr("dayu.cli.commands.interactive.Log.error", lambda message, **_kwargs: error_logs.append(str(message)))
+    monkeypatch.setattr("dayu.cli.commands.interactive.interactive", lambda *_args, **_kwargs: pytest.fail("不应进入 interactive"))
 
-    assert main() == 1
+    assert run_interactive_command(args) == 1
     assert any("interactive 单实例锁" in message for message in error_logs)
 
 
@@ -2318,21 +2416,24 @@ def test_main_prompt_path_returns_prompt_exit_code(monkeypatch: pytest.MonkeyPat
         prompt_kwargs.update(kwargs)
         return 7
 
-    monkeypatch.setattr("dayu.cli.main.parse_arguments", partial(_return_value, args))
-    monkeypatch.setattr("dayu.cli.main.setup_paths", partial(_return_value, workspace_config))
-    monkeypatch.setattr("dayu.cli.main.setup_model_name", partial(_return_value, model_name))
+    monkeypatch.setattr("dayu.cli.commands.prompt.setup_loglevel", lambda _args: None)
+    monkeypatch.setattr("dayu.cli.commands.prompt.setup_paths", partial(_return_value, workspace_config))
+    monkeypatch.setattr(
+        "dayu.cli.commands.prompt._build_execution_options",
+        lambda _args: SimpleNamespace(model_name="deepseek-thinking"),
+    )
     fake_dependencies = _FakeCliHostDependencies(
         running_config=running_config,
         scene_model_names={"prompt": "scene-prompt-model"},
     )
     monkeypatch.setattr(
-        "dayu.cli.main._prepare_cli_host_dependencies",
+        "dayu.cli.commands.prompt._prepare_cli_host_dependencies",
         lambda **_kwargs: fake_dependencies.as_tuple(),
     )
-    monkeypatch.setattr("dayu.cli.main._build_prompt_service", lambda **_kwargs: object())
-    monkeypatch.setattr("dayu.cli.main.prompt_command", _capture_prompt)
+    monkeypatch.setattr("dayu.cli.commands.prompt._build_prompt_service", lambda **_kwargs: object())
+    monkeypatch.setattr("dayu.cli.commands.prompt.prompt_command", _capture_prompt)
 
-    assert main() == 7
+    assert run_prompt_command(args) == 7
     prompt_execution_options = cast(Any, prompt_kwargs["execution_options"])
     assert prompt_kwargs["ticker"] == "AAPL"
     assert prompt_kwargs["show_thinking"] is False
@@ -2476,21 +2577,24 @@ def test_main_prompt_path_allows_missing_filings_dir(monkeypatch: pytest.MonkeyP
         prompt_kwargs.update(kwargs)
         return 9
 
-    monkeypatch.setattr("dayu.cli.main.parse_arguments", partial(_return_value, args))
-    monkeypatch.setattr("dayu.cli.main.setup_paths", partial(_return_value, workspace_config))
-    monkeypatch.setattr("dayu.cli.main.setup_model_name", partial(_return_value, model_name))
+    monkeypatch.setattr("dayu.cli.commands.prompt.setup_loglevel", lambda _args: None)
+    monkeypatch.setattr("dayu.cli.commands.prompt.setup_paths", partial(_return_value, workspace_config))
+    monkeypatch.setattr(
+        "dayu.cli.commands.prompt._build_execution_options",
+        lambda _args: SimpleNamespace(model_name="deepseek-thinking"),
+    )
     fake_dependencies = _FakeCliHostDependencies(
         running_config=running_config,
         scene_model_names={"prompt": "scene-prompt-model"},
     )
     monkeypatch.setattr(
-        "dayu.cli.main._prepare_cli_host_dependencies",
+        "dayu.cli.commands.prompt._prepare_cli_host_dependencies",
         lambda **_kwargs: fake_dependencies.as_tuple(),
     )
-    monkeypatch.setattr("dayu.cli.main._build_prompt_service", lambda **_kwargs: object())
-    monkeypatch.setattr("dayu.cli.main.prompt_command", _capture_prompt)
+    monkeypatch.setattr("dayu.cli.commands.prompt._build_prompt_service", lambda **_kwargs: object())
+    monkeypatch.setattr("dayu.cli.commands.prompt.prompt_command", _capture_prompt)
 
-    assert main() == 9
+    assert run_prompt_command(args) == 9
     assert prompt_kwargs["ticker"] == "AAPL"
 
 
@@ -2567,9 +2671,9 @@ def test_main_prompt_path_propagates_cli_options_to_mock_agent(
     monkeypatch.setattr("dayu.engine.toolset_registrars.register_web_toolset", web_tools_recorder)
     monkeypatch.setattr("dayu.engine.toolset_registrars.register_doc_toolset", doc_tools_recorder)
     monkeypatch.setattr("dayu.fins.toolset_registrars.register_fins_read_toolset", fins_read_tools_recorder)
-    monkeypatch.setattr("dayu.cli.main.Log.info", lambda *_args, **_kwargs: None)
-    monkeypatch.setattr("dayu.cli.main.Log.warning", lambda *_args, **_kwargs: None)
-    monkeypatch.setattr("dayu.cli.main.Log.error", lambda *_args, **_kwargs: None)
+    monkeypatch.setattr("dayu.cli.commands.prompt.Log.info", lambda *_args, **_kwargs: None)
+    monkeypatch.setattr("dayu.cli.commands.prompt.Log.warning", lambda *_args, **_kwargs: None)
+    monkeypatch.setattr("dayu.cli.commands.prompt.Log.error", lambda *_args, **_kwargs: None)
 
     assert main() == 0
     assert recorder.agent_create_args is not None
@@ -2708,9 +2812,9 @@ def test_main_prompt_path_sparse_override_preserves_run_json_base(
     monkeypatch.setattr("dayu.engine.toolset_registrars.register_web_toolset", web_tools_recorder)
     monkeypatch.setattr("dayu.engine.toolset_registrars.register_doc_toolset", doc_tools_recorder)
     monkeypatch.setattr("dayu.fins.toolset_registrars.register_fins_read_toolset", fins_read_tools_recorder)
-    monkeypatch.setattr("dayu.cli.main.Log.info", lambda *_args, **_kwargs: None)
-    monkeypatch.setattr("dayu.cli.main.Log.warning", lambda *_args, **_kwargs: None)
-    monkeypatch.setattr("dayu.cli.main.Log.error", lambda *_args, **_kwargs: None)
+    monkeypatch.setattr("dayu.cli.commands.prompt.Log.info", lambda *_args, **_kwargs: None)
+    monkeypatch.setattr("dayu.cli.commands.prompt.Log.warning", lambda *_args, **_kwargs: None)
+    monkeypatch.setattr("dayu.cli.commands.prompt.Log.error", lambda *_args, **_kwargs: None)
 
     assert main() == 0
 
@@ -2786,9 +2890,9 @@ def test_main_prompt_path_propagates_execution_permissions_to_web_tool_registrat
     monkeypatch.setattr("dayu.engine.toolset_registrars.register_web_toolset", web_tools_recorder)
     monkeypatch.setattr("dayu.engine.toolset_registrars.register_doc_toolset", doc_tools_recorder)
     monkeypatch.setattr("dayu.fins.toolset_registrars.register_fins_read_toolset", fins_read_tools_recorder)
-    monkeypatch.setattr("dayu.cli.main.Log.info", lambda *_args, **_kwargs: None)
-    monkeypatch.setattr("dayu.cli.main.Log.warning", lambda *_args, **_kwargs: None)
-    monkeypatch.setattr("dayu.cli.main.Log.error", lambda *_args, **_kwargs: None)
+    monkeypatch.setattr("dayu.cli.commands.prompt.Log.info", lambda *_args, **_kwargs: None)
+    monkeypatch.setattr("dayu.cli.commands.prompt.Log.warning", lambda *_args, **_kwargs: None)
+    monkeypatch.setattr("dayu.cli.commands.prompt.Log.error", lambda *_args, **_kwargs: None)
 
     assert main() == 0
     assert recorder.messages is not None
@@ -2915,9 +3019,9 @@ def test_main_prompt_path_propagates_run_json_defaults_to_host_and_agent(
     monkeypatch.setattr("dayu.engine.toolset_registrars.register_web_toolset", web_tools_recorder)
     monkeypatch.setattr("dayu.engine.toolset_registrars.register_doc_toolset", doc_tools_recorder)
     monkeypatch.setattr("dayu.fins.toolset_registrars.register_fins_read_toolset", fins_read_tools_recorder)
-    monkeypatch.setattr("dayu.cli.main.Log.info", lambda *_args, **_kwargs: None)
-    monkeypatch.setattr("dayu.cli.main.Log.warning", lambda *_args, **_kwargs: None)
-    monkeypatch.setattr("dayu.cli.main.Log.error", lambda *_args, **_kwargs: None)
+    monkeypatch.setattr("dayu.cli.commands.prompt.Log.info", lambda *_args, **_kwargs: None)
+    monkeypatch.setattr("dayu.cli.commands.prompt.Log.warning", lambda *_args, **_kwargs: None)
+    monkeypatch.setattr("dayu.cli.commands.prompt.Log.error", lambda *_args, **_kwargs: None)
 
     assert main() == 0
     assert recorder.agent_create_args is not None
@@ -3090,9 +3194,9 @@ def test_main_prompt_path_propagates_scene_manifest_prompt_assets_and_llm_model_
         lambda **_kwargs: "# SUBJECT_SLOT\n你正在分析的是 AAPL。",
     )
     monkeypatch.setattr("dayu.host.executor.build_async_agent", _MockPromptAgentBuilder(recorder))
-    monkeypatch.setattr("dayu.cli.main.Log.info", lambda *_args, **_kwargs: None)
-    monkeypatch.setattr("dayu.cli.main.Log.warning", lambda *_args, **_kwargs: None)
-    monkeypatch.setattr("dayu.cli.main.Log.error", lambda *_args, **_kwargs: None)
+    monkeypatch.setattr("dayu.cli.commands.prompt.Log.info", lambda *_args, **_kwargs: None)
+    monkeypatch.setattr("dayu.cli.commands.prompt.Log.warning", lambda *_args, **_kwargs: None)
+    monkeypatch.setattr("dayu.cli.commands.prompt.Log.error", lambda *_args, **_kwargs: None)
 
     assert main() == 0
     assert recorder.agent_create_args is not None
@@ -3154,8 +3258,6 @@ def test_main_non_interactive_path_returns_zero(monkeypatch: pytest.MonkeyPatch,
     args = Namespace(command=None, log_level=None, debug=False, verbose=False, info=False, quiet=False)
 
     monkeypatch.setattr("dayu.cli.main.parse_arguments", partial(_return_value, args))
-    monkeypatch.setattr("dayu.cli.main.setup_paths", partial(_return_value, workspace_config))
-    monkeypatch.setattr("dayu.cli.main.setup_model_name", partial(_return_value, model_name))
     assert main() == 0
 
 
@@ -3206,12 +3308,16 @@ def test_main_write_mode_requires_ticker(monkeypatch: pytest.MonkeyPatch, tmp_pa
         audit_model_name="deepseek-thinking",
     )
 
-    monkeypatch.setattr("dayu.cli.main.parse_arguments", partial(_return_value, args))
-    monkeypatch.setattr("dayu.cli.main.setup_paths", partial(_return_value, workspace_config))
-    monkeypatch.setattr("dayu.cli.main.setup_model_name", partial(_return_value, model_name))
-    monkeypatch.setattr("dayu.cli.main.Log.error", lambda *args, **kwargs: None)
+    monkeypatch.setattr("dayu.cli.commands.write.setup_loglevel", lambda _args: None)
+    monkeypatch.setattr("dayu.cli.commands.write.setup_paths", partial(_return_value, workspace_config))
+    monkeypatch.setattr("dayu.cli.commands.write.setup_model_name", partial(_return_value, model_name))
+    monkeypatch.setattr(
+        "dayu.cli.commands.write._build_execution_options",
+        lambda _args: SimpleNamespace(model_name="deepseek-thinking"),
+    )
+    monkeypatch.setattr("dayu.cli.commands.write.Log.error", lambda *args, **kwargs: None)
 
-    assert main() == 2
+    assert run_write_command(args) == 2
 
 
 @pytest.mark.unit
@@ -3263,12 +3369,16 @@ def test_main_write_summary_mode_requires_ticker(monkeypatch: pytest.MonkeyPatch
     )
 
     collector = _CallCollector()
-    monkeypatch.setattr("dayu.cli.main.parse_arguments", partial(_return_value, args))
-    monkeypatch.setattr("dayu.cli.main.setup_paths", partial(_return_value, workspace_config))
-    monkeypatch.setattr("dayu.cli.main.setup_model_name", partial(_return_value, model_name))
-    monkeypatch.setattr("dayu.cli.main.Log.error", collector.capture_error)
+    monkeypatch.setattr("dayu.cli.commands.write.setup_loglevel", lambda _args: None)
+    monkeypatch.setattr("dayu.cli.commands.write.setup_paths", partial(_return_value, workspace_config))
+    monkeypatch.setattr("dayu.cli.commands.write.setup_model_name", partial(_return_value, model_name))
+    monkeypatch.setattr(
+        "dayu.cli.commands.write._build_execution_options",
+        lambda _args: SimpleNamespace(model_name="deepseek-thinking"),
+    )
+    monkeypatch.setattr("dayu.cli.commands.write.Log.error", collector.capture_error)
 
-    assert main() == 2
+    assert run_write_command(args) == 2
     assert any("write --summary 模式要求必须提供 --ticker" in item for item in collector.error_logs)
 
 
@@ -3344,13 +3454,17 @@ def test_main_write_summary_mode_calls_print_report(monkeypatch: pytest.MonkeyPa
             captured_output_dir["value"] = output_dir
             return 6
 
-    monkeypatch.setattr("dayu.cli.main.parse_arguments", partial(_return_value, args))
-    monkeypatch.setattr("dayu.cli.main.setup_paths", partial(_return_value, workspace_config))
-    monkeypatch.setattr("dayu.cli.main.setup_model_name", partial(_return_value, model_name))
-    monkeypatch.setattr("dayu.cli.main.WriteService.print_report", _FakeWriteService.print_report)
-    monkeypatch.setattr("dayu.cli.main.run_write_pipeline", lambda **_kwargs: pytest.fail("summary 分支不应进入写作流水线"))
+    monkeypatch.setattr("dayu.cli.commands.write.setup_loglevel", lambda _args: None)
+    monkeypatch.setattr("dayu.cli.commands.write.setup_paths", partial(_return_value, workspace_config))
+    monkeypatch.setattr("dayu.cli.commands.write.setup_model_name", partial(_return_value, model_name))
+    monkeypatch.setattr(
+        "dayu.cli.commands.write._build_execution_options",
+        lambda _args: SimpleNamespace(model_name="deepseek-thinking"),
+    )
+    monkeypatch.setattr("dayu.cli.commands.write.WriteService.print_report", _FakeWriteService.print_report)
+    monkeypatch.setattr("dayu.cli.commands.write.run_write_pipeline", lambda **_kwargs: pytest.fail("summary 分支不应进入写作流水线"))
 
-    assert main() == 6
+    assert run_write_command(args) == 6
     assert captured_output_dir["value"] == (tmp_path / "draft" / "AAPL").resolve()
 
 
@@ -3404,20 +3518,24 @@ def test_main_write_mode_calls_pipeline(monkeypatch: pytest.MonkeyPatch, tmp_pat
     )
 
     collector = _CallCollector()
-    monkeypatch.setattr("dayu.cli.main.parse_arguments", partial(_return_value, args))
-    monkeypatch.setattr("dayu.cli.main.setup_paths", partial(_return_value, workspace_config))
-    monkeypatch.setattr("dayu.cli.main.setup_model_name", partial(_return_value, model_name))
+    monkeypatch.setattr("dayu.cli.commands.write.setup_loglevel", lambda _args: None)
+    monkeypatch.setattr("dayu.cli.commands.write.setup_paths", partial(_return_value, workspace_config))
+    monkeypatch.setattr("dayu.cli.commands.write.setup_model_name", partial(_return_value, model_name))
+    monkeypatch.setattr(
+        "dayu.cli.commands.write._build_execution_options",
+        lambda _args: SimpleNamespace(model_name="deepseek-thinking"),
+    )
     fake_dependencies = _FakeCliHostDependencies(running_config=running_config)
     monkeypatch.setattr(
-        "dayu.cli.main._prepare_cli_host_dependencies",
+        "dayu.cli.commands.write._prepare_cli_host_dependencies",
         lambda **_kwargs: fake_dependencies.as_tuple(),
     )
-    monkeypatch.setattr("dayu.cli.main._build_write_service", lambda **_kwargs: object())
-    monkeypatch.setattr("dayu.cli.main.Log.info", collector.capture_info)
-    monkeypatch.setattr("dayu.cli.main.Log.warn", collector.capture_warn)
-    monkeypatch.setattr("dayu.cli.main.run_write_pipeline", lambda **_kwargs: 4)
+    monkeypatch.setattr("dayu.cli.commands.write._build_write_service", lambda **_kwargs: object())
+    monkeypatch.setattr("dayu.cli.commands.write.Log.info", collector.capture_info)
+    monkeypatch.setattr("dayu.cli.commands.write.Log.warn", collector.capture_warn)
+    monkeypatch.setattr("dayu.cli.commands.write.run_write_pipeline", lambda **_kwargs: 4)
 
-    assert main() == 4
+    assert run_write_command(args) == 4
     assert any("写作流水线启动" in item for item in collector.info_logs)
     assert any("写作流水线结束但返回非零" in item for item in collector.warn_logs)
 
@@ -3474,20 +3592,24 @@ def test_main_write_mode_logs_success_when_pipeline_returns_zero(
     )
 
     collector = _CallCollector()
-    monkeypatch.setattr("dayu.cli.main.parse_arguments", partial(_return_value, args))
-    monkeypatch.setattr("dayu.cli.main.setup_paths", partial(_return_value, workspace_config))
-    monkeypatch.setattr("dayu.cli.main.setup_model_name", partial(_return_value, model_name))
+    monkeypatch.setattr("dayu.cli.commands.write.setup_loglevel", lambda _args: None)
+    monkeypatch.setattr("dayu.cli.commands.write.setup_paths", partial(_return_value, workspace_config))
+    monkeypatch.setattr("dayu.cli.commands.write.setup_model_name", partial(_return_value, model_name))
+    monkeypatch.setattr(
+        "dayu.cli.commands.write._build_execution_options",
+        lambda _args: SimpleNamespace(model_name="deepseek-thinking"),
+    )
     fake_dependencies = _FakeCliHostDependencies(running_config=running_config)
     monkeypatch.setattr(
-        "dayu.cli.main._prepare_cli_host_dependencies",
+        "dayu.cli.commands.write._prepare_cli_host_dependencies",
         lambda **_kwargs: fake_dependencies.as_tuple(),
     )
-    monkeypatch.setattr("dayu.cli.main._build_write_service", lambda **_kwargs: object())
-    monkeypatch.setattr("dayu.cli.main.Log.info", collector.capture_info)
-    monkeypatch.setattr("dayu.cli.main.Log.warn", collector.capture_warn)
-    monkeypatch.setattr("dayu.cli.main.run_write_pipeline", lambda **_kwargs: 0)
+    monkeypatch.setattr("dayu.cli.commands.write._build_write_service", lambda **_kwargs: object())
+    monkeypatch.setattr("dayu.cli.commands.write.Log.info", collector.capture_info)
+    monkeypatch.setattr("dayu.cli.commands.write.Log.warn", collector.capture_warn)
+    monkeypatch.setattr("dayu.cli.commands.write.run_write_pipeline", lambda **_kwargs: 0)
 
-    assert main() == 0
+    assert run_write_command(args) == 0
     assert any("写作流水线完成: exit_code=0" in item for item in collector.info_logs)
     assert not collector.warn_logs
 
@@ -3545,20 +3667,24 @@ def test_main_write_mode_logs_elapsed_when_pipeline_raises(
 
     collector = _CallCollector()
 
-    monkeypatch.setattr("dayu.cli.main.parse_arguments", partial(_return_value, args))
-    monkeypatch.setattr("dayu.cli.main.setup_paths", partial(_return_value, workspace_config))
-    monkeypatch.setattr("dayu.cli.main.setup_model_name", partial(_return_value, model_name))
+    monkeypatch.setattr("dayu.cli.commands.write.setup_loglevel", lambda _args: None)
+    monkeypatch.setattr("dayu.cli.commands.write.setup_paths", partial(_return_value, workspace_config))
+    monkeypatch.setattr("dayu.cli.commands.write.setup_model_name", partial(_return_value, model_name))
+    monkeypatch.setattr(
+        "dayu.cli.commands.write._build_execution_options",
+        lambda _args: SimpleNamespace(model_name="deepseek-thinking"),
+    )
     fake_dependencies = _FakeCliHostDependencies(running_config=running_config)
     monkeypatch.setattr(
-        "dayu.cli.main._prepare_cli_host_dependencies",
+        "dayu.cli.commands.write._prepare_cli_host_dependencies",
         lambda **_kwargs: fake_dependencies.as_tuple(),
     )
-    monkeypatch.setattr("dayu.cli.main._build_write_service", lambda **_kwargs: object())
-    monkeypatch.setattr("dayu.cli.main.Log.info", collector.capture_info)
-    monkeypatch.setattr("dayu.cli.main.Log.error", collector.capture_error)
-    monkeypatch.setattr("dayu.cli.main.run_write_pipeline", _raise_runtime_error_for_write_pipeline)
+    monkeypatch.setattr("dayu.cli.commands.write._build_write_service", lambda **_kwargs: object())
+    monkeypatch.setattr("dayu.cli.commands.write.Log.info", collector.capture_info)
+    monkeypatch.setattr("dayu.cli.commands.write.Log.error", collector.capture_error)
+    monkeypatch.setattr("dayu.cli.commands.write.run_write_pipeline", _raise_runtime_error_for_write_pipeline)
 
-    assert main() == 2
+    assert run_write_command(args) == 2
     assert any("写作模式执行失败: elapsed=" in item for item in collector.error_logs)
 
 
@@ -3605,12 +3731,16 @@ def test_main_write_returns_130_when_run_write_pipeline_is_cancelled(
 
     collector = _CallCollector()
 
-    monkeypatch.setattr("dayu.cli.main.parse_arguments", partial(_return_value, args))
-    monkeypatch.setattr("dayu.cli.main.setup_paths", partial(_return_value, workspace_config))
-    monkeypatch.setattr("dayu.cli.main.setup_model_name", partial(_return_value, model_name))
+    monkeypatch.setattr("dayu.cli.commands.write.setup_loglevel", lambda _args: None)
+    monkeypatch.setattr("dayu.cli.commands.write.setup_paths", partial(_return_value, workspace_config))
+    monkeypatch.setattr("dayu.cli.commands.write.setup_model_name", partial(_return_value, model_name))
+    monkeypatch.setattr(
+        "dayu.cli.commands.write._build_execution_options",
+        lambda _args: SimpleNamespace(model_name="deepseek-thinking"),
+    )
     fake_dependencies = _FakeCliHostDependencies(running_config=running_config)
     monkeypatch.setattr(
-        "dayu.cli.main._prepare_cli_host_dependencies",
+        "dayu.cli.commands.write._prepare_cli_host_dependencies",
         lambda **_kwargs: fake_dependencies.as_tuple(),
     )
 
@@ -3638,18 +3768,18 @@ def test_main_write_returns_130_when_run_write_pipeline_is_cancelled(
             return cancel_callback()
 
     monkeypatch.setattr(
-        "dayu.cli.main._build_write_service",
+        "dayu.cli.commands.write._build_write_service",
         lambda **_kwargs: WriteService(
             host=cast(HostedExecutionGatewayProtocol, _CancellingWriteHost()),
             workspace=cast(Any, fake_dependencies.workspace),
             scene_execution_acceptance_preparer=cast(Any, fake_dependencies.scene_execution_acceptance_preparer),
         ),
     )
-    monkeypatch.setattr("dayu.cli.main.Log.info", collector.capture_info)
-    monkeypatch.setattr("dayu.cli.main.Log.warn", collector.capture_warn)
-    monkeypatch.setattr("dayu.cli.main.Log.error", collector.capture_error)
+    monkeypatch.setattr("dayu.cli.commands.write.Log.info", collector.capture_info)
+    monkeypatch.setattr("dayu.cli.commands.write.Log.warn", collector.capture_warn)
+    monkeypatch.setattr("dayu.cli.commands.write.Log.error", collector.capture_error)
 
-    assert main() == 130
+    assert run_write_command(args) == 130
     assert any("写作模式已取消: exit_code=130" in item for item in collector.warn_logs)
     assert not any("写作流水线结束但返回非零" in item for item in collector.warn_logs)
 
@@ -3694,12 +3824,12 @@ def test_main_fins_command_bypasses_setup_paths(monkeypatch: pytest.MonkeyPatch)
     )
 
     monkeypatch.setattr("dayu.cli.main.parse_arguments", partial(_return_value, args))
-    monkeypatch.setattr("dayu.cli.main._run_fins_command", lambda _args: 0)
+    monkeypatch.setattr("dayu.cli.commands.fins.run_fins_command", lambda _args: 0)
 
     def _fail_setup_paths(_args: object) -> object:
         raise AssertionError("fins 命令不应调用 setup_paths")
 
-    monkeypatch.setattr("dayu.cli.main.setup_paths", _fail_setup_paths)
+    monkeypatch.setattr("dayu.cli.commands.interactive.setup_paths", _fail_setup_paths)
 
     assert main() == 0
 
@@ -3713,7 +3843,7 @@ def test_build_fins_command_normalizes_download_payload(monkeypatch: pytest.Monk
         namespace.ticker = "BABA"
         namespace.ticker_aliases = ["9988", "9988.HK"]
 
-    monkeypatch.setattr("dayu.cli.fins_commands.prepare_cli_args", _fake_prepare_cli_args)
+    monkeypatch.setattr("dayu.cli.commands.fins.prepare_cli_args", _fake_prepare_cli_args)
 
     command = _build_fins_command(
         Namespace(
@@ -3753,7 +3883,7 @@ def test_build_fins_command_preserves_upload_filing_infer_results(
         namespace.company_name = "Apple Inc."
         namespace.ticker_aliases = ["AAPL.US"]
 
-    monkeypatch.setattr("dayu.cli.fins_commands.prepare_cli_args", _fake_prepare_cli_args)
+    monkeypatch.setattr("dayu.cli.commands.fins.prepare_cli_args", _fake_prepare_cli_args)
 
     command = _build_fins_command(
         Namespace(
@@ -3853,7 +3983,7 @@ def test_build_fins_command_allows_upload_material_delete_without_files(tmp_path
 
 @pytest.mark.unit
 def test_run_fins_command_stream_path(monkeypatch: pytest.MonkeyPatch, tmp_path: Path, capsys: pytest.CaptureFixture[str]) -> None:
-    """验证 `_run_fins_command` 可消费流式结果并输出最终摘要。"""
+    """验证 `run_fins_command` 可消费流式结果并输出最终摘要。"""
 
     class _FakeService:
         def __init__(self) -> None:
@@ -3884,8 +4014,8 @@ def test_run_fins_command_stream_path(monkeypatch: pytest.MonkeyPatch, tmp_path:
 
             return FinsSubmission(session_id="test-session", execution=_stream())
 
-    monkeypatch.setattr("dayu.cli.fins_commands._build_fins_ops_service", lambda _args: _FakeService())
-    monkeypatch.setattr("dayu.cli.fins_commands.format_fins_cli_result", lambda command, result: f"{command.value}:{result.status}")
+    monkeypatch.setattr("dayu.cli.commands.fins._build_fins_ops_service", lambda _args: _FakeService())
+    monkeypatch.setattr("dayu.cli.commands.fins.format_fins_cli_result", lambda command, result: f"{command.value}:{result.status}")
 
     args = Namespace(
         command="download",
@@ -3903,7 +4033,7 @@ def test_run_fins_command_stream_path(monkeypatch: pytest.MonkeyPatch, tmp_path:
         quiet=False,
     )
 
-    assert __import__("dayu.cli.main", fromlist=["dummy"])._run_fins_command(args) == 0
+    assert run_fins_command(args) == 0
     output = capsys.readouterr().out
     assert "download:ok" in output
 
@@ -3965,10 +4095,10 @@ def test_run_fins_command_upload_stream_logs_progress_at_info(
         del module
         verbose_lines.append(message)
 
-    monkeypatch.setattr("dayu.cli.fins_commands._build_fins_ops_service", lambda _args: _FakeService())
-    monkeypatch.setattr("dayu.cli.fins_commands.format_fins_cli_result", lambda command, result: f"{command.value}:{result.status}")
-    monkeypatch.setattr("dayu.cli.main.Log.info", _capture_info)
-    monkeypatch.setattr("dayu.cli.main.Log.verbose", _capture_verbose)
+    monkeypatch.setattr("dayu.cli.commands.fins._build_fins_ops_service", lambda _args: _FakeService())
+    monkeypatch.setattr("dayu.cli.commands.fins.format_fins_cli_result", lambda command, result: f"{command.value}:{result.status}")
+    monkeypatch.setattr("dayu.cli.commands.fins.Log.info", _capture_info)
+    monkeypatch.setattr("dayu.cli.commands.fins.Log.verbose", _capture_verbose)
 
     args = Namespace(
         command="upload_filing",
@@ -3991,7 +4121,7 @@ def test_run_fins_command_upload_stream_logs_progress_at_info(
         quiet=False,
     )
 
-    assert __import__("dayu.cli.main", fromlist=["dummy"])._run_fins_command(args) == 0
+    assert run_fins_command(args) == 0
     output = capsys.readouterr().out
     assert "upload_filing:ok" in output
     assert any("[upload_filing] upload_started ticker=AAPL action=create file_count=1" in line for line in info_lines)
@@ -4051,8 +4181,8 @@ def test_run_fins_command_download_allows_missing_filings_dir(
 
             return FinsSubmission(session_id="test-session", execution=_stream())
 
-    monkeypatch.setattr("dayu.cli.fins_commands._build_fins_ops_service", lambda _args: _FakeService())
-    monkeypatch.setattr("dayu.cli.fins_commands.format_fins_cli_result", lambda command, result: f"{command.value}:{result.status}")
+    monkeypatch.setattr("dayu.cli.commands.fins._build_fins_ops_service", lambda _args: _FakeService())
+    monkeypatch.setattr("dayu.cli.commands.fins.format_fins_cli_result", lambda command, result: f"{command.value}:{result.status}")
 
     args = Namespace(
         command="download",
@@ -4071,6 +4201,6 @@ def test_run_fins_command_download_allows_missing_filings_dir(
         quiet=False,
     )
 
-    assert __import__("dayu.cli.main", fromlist=["dummy"])._run_fins_command(args) == 0
+    assert run_fins_command(args) == 0
     output = capsys.readouterr().out
     assert "download:ok" in output
