+++ tests/engine/test_config_loader_coverage.py
@@ -10,7 +10,7 @@ from typing import Any, cast
 import pytest
 
 from dayu.startup import config_loader as module
-from dayu.startup.config_file_resolver import ConfigFileResolver, _resolve_package_config_path
+from dayu.startup.config_file_resolver import ConfigFileResolver, resolve_package_config_path
 
 
 @pytest.mark.unit
@@ -54,7 +54,7 @@ def test_config_loader_init_branches(monkeypatch: pytest.MonkeyPatch, tmp_path: 
     loader_default = module.ConfigLoader(ConfigFileResolver())
     assert len(loader_default._resolver.config_dirs) == 1
 
-    package_config_path = _resolve_package_config_path()
+    package_config_path = resolve_package_config_path()
     loader_same = module.ConfigLoader(ConfigFileResolver(package_config_path))
     assert loader_same._resolver.config_dirs == [package_config_path]
 
@@ -88,7 +88,7 @@ def test_config_file_resolver_read_text_required_and_optional_branches(
     assert resolver.read_text("optional_missing.json", required=False) is None
     assert warnings
 
-    app_config_resolver = ConfigFileResolver(_resolve_package_config_path())
+    app_config_resolver = ConfigFileResolver(resolve_package_config_path())
     with pytest.raises(FileNotFoundError, match="查找路径"):
         app_config_resolver.read_text("also_missing.json", required=True)
 
@@ -97,7 +97,7 @@ def test_config_file_resolver_read_text_required_and_optional_branches(
 def test_config_file_resolver_read_json_optional_none_and_decode_error(monkeypatch: pytest.MonkeyPatch) -> None:
     """验证 ConfigFileResolver JSON 解析的 None 与失败分支。"""
 
-    resolver = ConfigFileResolver(_resolve_package_config_path())
+    resolver = ConfigFileResolver(resolve_package_config_path())
 
     monkeypatch.setattr(resolver, "read_text", lambda filename, required=True: None)
     assert resolver.read_json("x.json", required=False) is None
@@ -217,7 +217,7 @@ def test_collect_model_referenced_env_vars_only_scans_selected_models(tmp_path: 
 def test_load_run_config_rejects_non_object(monkeypatch: pytest.MonkeyPatch) -> None:
     """验证 run.json 顶层不是对象时会显式失败。"""
 
-    resolver = ConfigFileResolver(_resolve_package_config_path())
+    resolver = ConfigFileResolver(resolve_package_config_path())
     monkeypatch.setattr(resolver, "read_json", lambda filename, required=True: ["bad"])
 
     loader = module.ConfigLoader(resolver)
@@ -230,7 +230,7 @@ def test_load_run_config_rejects_non_object(monkeypatch: pytest.MonkeyPatch) -> 
 def test_load_llm_models_rejects_non_object_model_entry(monkeypatch: pytest.MonkeyPatch) -> None:
     """验证 llm_models.json 中单个模型配置不是对象时会显式失败。"""
 
-    resolver = ConfigFileResolver(_resolve_package_config_path())
+    resolver = ConfigFileResolver(resolve_package_config_path())
     monkeypatch.setattr(
         resolver,
         "read_json",
@@ -247,7 +247,7 @@ def test_load_llm_models_rejects_non_object_model_entry(monkeypatch: pytest.Monk
 def test_load_llm_models_ignores_metadata_keys(monkeypatch: pytest.MonkeyPatch) -> None:
     """验证 llm_models.json 顶层元信息键不会被当成模型配置。"""
 
-    resolver = ConfigFileResolver(_resolve_package_config_path())
+    resolver = ConfigFileResolver(resolve_package_config_path())
     monkeypatch.setattr(
         resolver,
         "read_json",
@@ -286,7 +286,7 @@ def test_extract_env_var_names_and_should_scan_helpers(tmp_path: Path) -> None:
 def test_load_llm_model_replaces_env_vars_and_reports_missing_model(monkeypatch: pytest.MonkeyPatch) -> None:
     """加载单模型时应替换环境变量，并在模型不存在时报错。"""
 
-    resolver = ConfigFileResolver(_resolve_package_config_path())
+    resolver = ConfigFileResolver(resolve_package_config_path())
     monkeypatch.setenv("TEST_API_KEY", "secret-token")
     monkeypatch.setattr(
         resolver,
@@ -320,7 +320,7 @@ def test_load_llm_model_replaces_env_vars_and_reports_missing_model(monkeypatch:
 def test_load_toolset_registrars_caches_and_validates_entries(monkeypatch: pytest.MonkeyPatch) -> None:
     """toolset registrar 加载应返回副本，并校验 key/value。"""
 
-    resolver = ConfigFileResolver(_resolve_package_config_path())
+    resolver = ConfigFileResolver(resolve_package_config_path())
     payload = {"doc": "dayu.engine.registrar:register"}
     monkeypatch.setattr(resolver, "read_json", lambda filename, required=True: payload)
